refactor: seperated styles from components
